Alteração na funcionalidade de like

+++ backend/src/controllers/LikeController.js
@@ -17,16 +17,22 @@ class LikeController{
         const authHeader = req.headers.authorization;
         const [,token] = authHeader.split(" ");
         const {id} = await promisify(jwt.verify)(token,process.env.JWT_SECRET); //Seu id
+        const post = await Posts.findOne({where:{id:post_id}});
+
+        if(!post) return res.status(400).json({msg:"Postagem não existe!"});
         
         const {username} = await Users.findOne({where:{id}});
-        
+
         const verifyLike = await Likes.findOne({where:{user_id:id, post_id}});
-        if(verifyLike){
+
+        if(verifyLike)
+        {
             const dislike = await Likes.destroy({where:{user_id:id, post_id, username}});
             return res.status(200).json({msg:"+1 👎"});
         }
 
         const like = await Likes.create({user_id:id, post_id, username});
+
         return res.status(203).json({msg:"+1 👍"});
 
         } catch (error) {
